Grows Partitions too * Depends on /opt/scripts/tools/grow_partitions.sh * Removed definitions from src/utils.hpp to src/utils.cpp
Demo:
https://asciinema.org/a/427430
https://asciinema.org/a/427432
