+++ CMakeLists.txt
@@ -32,6 +32,7 @@ add_executable(beagle-config
   src/ui/ui.cpp
   src/ui/ui.hpp
   src/utils.hpp
+  src/utils.cpp
 )
 
 target_link_libraries(beagle-config
